requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
feedgen>=0.9.0
pytz>=2023.3
urllib3>=2.0.0 